WARN: Type cast to avoid signed/unsigned warning.

+++ src/metaImage.cxx
@@ -767,7 +767,8 @@ ElementByteOrderSwap(METAIO_STL::streamoff _quantity)
       }
     case 2:
       {
-      for(size_t i=0; i<quantity*m_ElementNumberOfChannels; i++)
+      for(size_t i=0;
+        i<static_cast<size_t>(quantity*m_ElementNumberOfChannels); i++)
         {
         ((MET_USHORT_TYPE *)m_ElementData)[i] =
               MET_ByteOrderSwapShort(((MET_USHORT_TYPE *)m_ElementData)[i]);
@@ -776,7 +777,8 @@ ElementByteOrderSwap(METAIO_STL::streamoff _quantity)
       }
     case 4:
       {
-      for(size_t i=0; i<quantity*m_ElementNumberOfChannels; i++)
+      for(size_t i=0;
+        i<static_cast<size_t>(quantity*m_ElementNumberOfChannels); i++)
         {
         ((MET_UINT_TYPE *)m_ElementData)[i] =
               MET_ByteOrderSwapLong(((MET_UINT_TYPE *)m_ElementData)[i]);
@@ -786,7 +788,8 @@ ElementByteOrderSwap(METAIO_STL::streamoff _quantity)
     case 8:
       {
       char* data = (char*)m_ElementData;
-      for(size_t i=0; i<quantity*m_ElementNumberOfChannels; i++)
+      for(size_t i=0;
+        i<static_cast<size_t>(quantity*m_ElementNumberOfChannels); i++)
         {
         MET_ByteOrderSwap8(data);
         data += 8;
@@ -836,7 +839,8 @@ ElementMinMaxRecalc(void)
   MET_ValueToDouble(m_ElementType, m_ElementData, 0, &tf);
   m_ElementMin = tf;
   m_ElementMax = tf;
-  for(size_t i=1; i<m_Quantity*m_ElementNumberOfChannels; i++)
+  for(size_t i=1;
+    i<static_cast<size_t>(m_Quantity*m_ElementNumberOfChannels); i++)
     {
     MET_ValueToDouble(m_ElementType, m_ElementData, i, &tf);
     if(tf<m_ElementMin)
@@ -991,7 +995,8 @@ ConvertElementDataTo(MET_ValueEnumType _elementType,
     ElementMinMaxRecalc();
     }
 
-  for(size_t i=0; i<m_Quantity*m_ElementNumberOfChannels; i++)
+  for(size_t i=0;
+    i<static_cast<size_t>(m_Quantity*m_ElementNumberOfChannels); i++)
     {
     MET_ValueToValue(m_ElementType, m_ElementData, i, _elementType,
                      newElementData, m_ElementMin, m_ElementMax,
